Issue #199: Fix incorrect return value from MIKMIDIClockSecondsPerMIDITimeStamp() on arm64

+++ Source/MIKMIDIClock.m
@@ -364,7 +364,7 @@ Float64 MIKMIDIClockSecondsPerMIDITimeStamp()
 	dispatch_once(&onceToken, ^{
 		mach_timebase_info_data_t timeBaseInfoData;
 		mach_timebase_info(&timeBaseInfoData);
-		secondsPerMIDITimeStamp = (timeBaseInfoData.numer / timeBaseInfoData.denom) / 1.0e9;
+		secondsPerMIDITimeStamp = ((Float64)timeBaseInfoData.numer / (Float64)timeBaseInfoData.denom) / 1.0e9;
 	});
 	return secondsPerMIDITimeStamp;
